aha-common-utils/pyproject.toml
[project]
name = "aha-common-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "geebytes", email = "vforfreedom96@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "orjson>=3.10.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.18",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "toml>=0.10.2",
    "typer>=0.21.1",
    "psycopg2>=2.9.11",
    "tomli>=2.4.0",
    "sentry-sdk[fastapi]>=2.19.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-logging>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0"
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
